feat: CQDG-1123 implement public studies page
