+++ src/views/Studies/components/PageContent/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import intl from 'react-intl-universal';
 import { useDispatch } from 'react-redux';
-import { Link } from 'react-router-dom';
+import { Link, useLocation } from 'react-router-dom';
 import { ProfileOutlined } from '@ant-design/icons';
 import ProLabel from '@ferlab/ui/core/components/ProLabel';
 import ProTable from '@ferlab/ui/core/components/ProTable';
@@ -154,6 +154,9 @@ const PageContent = ({ defaultColumns = [] }: OwnProps) => {
   );
   const isProgramsEnabled: boolean = EnvVariables.configFor('PROGRAMS_ENABLED') === 'true';
 
+  const location = useLocation();
+  const isPublicStudiesPage = location.pathname === STATIC_ROUTES.PUBLIC_STUDIES;
+
   const { loading, total, data } = useStudies({
     first: PAGE_SIZE,
     offset: PAGE_SIZE * (queryConfig.pageIndex - 1),
@@ -239,12 +242,12 @@ const PageContent = ({ defaultColumns = [] }: OwnProps) => {
                 pageSize: PAGE_SIZE,
                 total,
               },
-              enableColumnSort: true,
+              enableColumnSort: !isPublicStudiesPage,
               onColumnSortChange: (newState) =>
                 dispatch(
                   updateUserConfig({ studies: { tables: { studies: { columns: newState } } } }),
                 ),
-              enableTableExport: true,
+              enableTableExport: !isPublicStudiesPage,
               onTableExportClick: () => {
                 dispatch(
                   fetchTsvReport({
